Create Abstract
Smart Gloves Using Arduino Uno is a social purpose project in which we try to implement a system that makes the communication gap between deaf people and abilities as small as possible. The Deaf-Mute People use sign language or gestures to communicate, but it is impossible to understand capable people when the Arduino UNO board is used as a control board to connect all the sensors (Flex Sensors & Bluetooth Module). This interacting device is a microcontroller-based system, which is the outline for reducing the communication space between deaf-mutes and normal people. This system can be accordingly configured to work as a "smart device". The Arduino Uno (At-mega 328 microcontroller), Bluetooth Module, application, and flex sensors are used in this project. The device under consideration is a hybrid of a glove and a microcontroller-based system. Data gloves are used to detect hand movements, which are then translated into human speech by a microcontroller-based system. The data glove is furnished with four flex sensors placed on the glove. Deaf-mute people's hand movements are converted into a text and a speech signal by the date glove worn on their hands, making this system useful for them.